+++ chord_metadata_service/chord/export/cbioportal.py
@@ -5,7 +5,7 @@ import re
 
 from django.db.models import F
 
-from .export_utils import ExportError
+from .utils import ExportError
 
 from chord_metadata_service.chord.models import Dataset
 from chord_metadata_service.patients.models import Individual
