+++ chord_metadata_service/chord/export/metadata.py
@@ -1,11 +1,19 @@
-import logging
-
 from chord_metadata_service.chord.models import Dataset, Project, Table
 from chord_metadata_service.chord.workflows.metadata import WORKFLOW_CBIOPORTAL
 
-from .export_cbio import study_export as export_cbioportal_workflow
+from .cbioportal import study_export as export_cbioportal_workflow
 
-logger = logging.getLogger(__name__)
+__all__ = [
+    "OBJECT_TYPE_PROJECT",
+    "OBJECT_TYPE_DATASET",
+    "OBJECT_TYPE_TABLE",
+
+    "EXPORT_OBJECT_TYPE",
+    "EXPORT_FORMATS",
+    "EXPORT_FORMAT_FUNCTION_MAP",
+    "EXPORT_FORMAT_OBJECT_TYPE_MAP",
+]
+
 
 OBJECT_TYPE_PROJECT = "project"
 OBJECT_TYPE_DATASET = "dataset"
